+++ mapchete_hub/geometry.py
@@ -6,7 +6,7 @@ from mapchete.io import path_exists, absolute_path as mp_absolute_path
 from mapchete.io.vector import reproject_geometry
 from mapchete.tile import BufferedTilePyramid
 from rasterio.crs import CRS
-from shapely import from_wkt
+from shapely import from_wkt, intersection
 from shapely.geometry import box, mapping, shape
 from shapely.ops import cascaded_union
 
@@ -58,8 +58,21 @@ def process_area_from_config(job_config: models.MapcheteJob, dst_crs=None, **kwa
         pixelbuffer=config["pyramid"].get("pixelbuffer", 0),
     )
 
-    # bounds
-    if params.get("bounds"):
+    # bounds and area
+    if params.get("bounds") and params.get("area"):
+        geometry_bounds = box(*params.get("bounds"))
+        if path_exists(params.get("area")):
+            absolute_path = mp_absolute_path(params.get("area"))
+            all_geoms = []
+            with fiona.open(str(absolute_path), mode="r") as src:
+                for s in src:
+                    all_geoms.append(shape(s['geometry']))
+            geometry_area = cascaded_union(all_geoms)
+        else:
+            geometry_area = from_wkt(params.get("area"))
+        geometry = intersection(geometry_bounds, geometry_area)
+    # bounds        
+    elif params.get("bounds"):
         geometry = box(*params.get("bounds"))
     # geometry
     elif params.get("geometry"):